pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "Reusable AI agent orchestration library for iterative document refinement"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "orchestration", "llm", "langchain", "refinement"]

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
]
google = [
    "langchain-google-genai>=2.0.0",
]
anthropic = [
    "langchain-anthropic>=0.3.0",
]
all-providers = [
    "ai-orchestrator[langchain,google,anthropic]",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "websockets>=10.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.0.0",
    "aiofiles>=23.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/va2ai/ai-agent-orchestration-platform"
Documentation = "https://github.com/va2ai/ai-agent-orchestration-platform#readme"
Repository = "https://github.com/va2ai/ai-agent-orchestration-platform.git"
Issues = "https://github.com/va2ai/ai-agent-orchestration-platform/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ai_orchestrator = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
